==> Instacart Basket Analysis/03 Scripts/1.1 IC Data Wrangling and Subsetting.ipynb <==
# 1.1 IC Data Wrangling and Subsetting

## Contents
### Import libraries and data sets
### Dropping and renaming columns
### Changing data type
### Transposing data
### Subsetting
### Wrangling orders data set
### Data Questions
### Export wrangled dataframes

## Import libraries and data sets

# import libraries
import pandas as pd
import numpy as np
import os

# create path to project folder
path = r'/Users/susanwang/Documents/CF_Tasks/Instacart Basket Analysis'

# import orders df
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# import products df
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

## Dropping and renaming columns

# drop column eval_set from orders df
df_ords = df_ords.drop(columns = ['eval_set'])

# check for missing values in days_since_prior_order column in orders df
df_ords['days_since_prior_order'].value_counts(dropna=False)

# renaming order_dow column in orders df
df_ords.rename(columns={'order_dow':'orders_day_of_the_week'}, inplace=True)

df_ords.head()

## Changing data type

#changing data type on order_id column
df_ords['order_id'] = df_ords['order_id'].astype('str')

# check data type of single column
df_ords['order_id'].dtype

## Transposing Data

# import data set departments.csv
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col=False)

df_dep.head()

# transposing df_dep
df_dep_t = df_dep.T

df_dep_t

# add index to df
df_dep_t.reset_index()

# create new header from first row
new_header = df_dep_t.iloc[0]

# take data after first row
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# add new header 
df_dep_t_new.columns = new_header

df_dep_t_new

# turn df in to data dictionary
data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

# see what dep id 19 is
print(data_dict.get('19'))

## Subsetting

# subsetting using the index operator
df_snacks = df_prods[df_prods['department_id']==19]

df_snacks

# subsetting using loc function
df_snacks_2 = df_prods.loc[df_prods['department_id']==19]

df_snacks_2.head()

# subsetting using loc function and list
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

df_snacks_3.head()

### Wrangling orders data set

df_ords.info()

# 2. change data type of identifier variable user_id
df_ords['user_id'] = df_ords['user_id'].astype('str')

# check data type
df_ords['user_id'].dtype

## Data questions

#### Your client wants to know what the busiest hour is for placing orders. Find the frequency of the corresponding variable and share your findings.

# Busiest hour for placing orders
df_ords['order_hour_of_day'].value_counts(dropna=True)

The busiest hour is at 10 am, followed by 11 am.

#### Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary

# what is 4 in dep id
print(data_dict.get('4'))

#### The sales team in your client’s organization wants to know more about breakfast item sales. Create a subset containing only the required information.

# breakfast items is dep_id 14
df_breakfast = df_prods.loc[df_prods['department_id']==14]

df_breakfast.head()

#### Details about products that customers might use to throw dinner parties. Find all observations from the entire dataframe that include items from the following departments: alcohol, deli, beverages, and meat/seafood. 

# use loc function with isin and list depts
# alcohol: 5, deli: 20, beverages: 7, meat/seafood: 12
df_dinner = df_prods.loc[df_prods['department_id'].isin([5, 7, 12, 20])]

df_dinner.head(20)

# 8. how many rows does df_dinner have
df_dinner.shape

#### Analyze the customer with a "user_id" of “1.”

# return a df with just this customer
df_user_1 = df_ords.loc[df_ords['user_id']=='1']

df_user_1

It seems this user 1 made all the first ten orders, separated by many days, and always between Sunday and Wednesday.

# What basic stats on this users behaviour can you provide based on the information you have?
df_user_1.describe()

## Export wrangled dataframes

# export df_ords
df_ords.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index=False)

# export df_dep_t_new
df_dep_t_new.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'departments_wrangled.csv'))
