==> Instacart Basket Analysis/03 Scripts/1.2 IC Data Consistency Checks.ipynb <==
# 1.2 IC Data Consistency Checks

## Contents
### Import libraries and data sets
### Missing values in products df
### Duplicates in products df
### Export cleaned products data set
### Data consistency checks on the orders data set
### Mixed type data in orders df
### Missing values in orders df
### Duplicates in orders df
### Export cleaned orders df


## Import libraries and data sets

import pandas as pd
import numpy as np
import os

# create path to project folder
path = r'/Users/susanwang/Documents/CF_Tasks/Instacart Basket Analysis'

# import products data set
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col=False)

df_prods.head()

# import orders data set
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col=False)

df_ords.head()

## Missing Values

# check df_prods for missing values
df_prods.isnull().sum()

# create dataframe with only null values
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

# df_prods shape to check rows
df_prods.shape

# create clean df dropping nan rows
df_prods_clean = df_prods[df_prods['product_name'].isnull()==False]

# check clean df shape
df_prods_clean.shape

## Duplicates

# create new df with duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

# check df shape rows
df_prods_clean.shape

# drop duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

#check new df shape
df_prods_clean_no_dups.shape

## Export cleaned products dataset

df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index=False)

## Data consistency checks on the orders dataset

# df_ords describe
df_ords.describe()

The 'days_since_prior_order' column has less count than the other columns. Are there NaN values there?
The minimums look fine. For the 'day_of_the_week', 0 means Saturday. For hour_of_day, it would be midnight. 
The maximum order_number is 100, but the 25% is 5, 50% is 11, and 75% is 23. We should check the order_number column.

## Mixed type data in orders df

# use for loop, applymap deprecated, using map instead
for col in df_ords.columns.tolist():
    weird = (df_ords[[col]].map(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df_ords[weird]) > 0:
        print (col)
    else: 
        print(f'all one type for {col}')


No mixed data found.

# checking dtypes
df_ords.dtypes

Order_id and user_id shouldn't be integers. Will need to go back and redo the wrangling step:

# change data type for order_id
df_ords['order_id'] = df_ords['order_id'].astype('str')

# change data type for user_id
df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords.dtypes

## Missing values in orders df

# finding using isnull
df_ords.isnull().sum()

There are 206209 missing values in days_since_prior_order. That is the same number as the maximum in the user_id column.
Every customer must make his first order, but there can't be a calulation between that first order and his prior order, so that is why the value is missing. 

# address missing values by creating a new variable
df_ords['first_order'] = df_ords['days_since_prior_order'].isnull()

There are too many missing values in this column, so removing the rows is not a good idea.

I cannot impute the average value, because that would distort the data. We want to keep the information about the customer's first order.
I also cannot impute a representative number, like -1. -1 could signal that this was the first order, but if I were to run aggregations on this column, the results would get distorted.

So the best choice is to create a new variable to flag these rows as the first order for the customers.

*Since this was not taught in the course yet, another option would be to leave the NaN's there, and make note about what it means. 

# check value counts
flag_count = df_ords['first_order'].value_counts(dropna=False)

flag_count

# export frequency table to excel report
flag_count.to_clipboard()

## Duplicates in the orders df

# check for duplicaates in a new df
df_ords_dups = df_ords[df_ords.duplicated()]

There seems to be no duplicates.

## Export cleaned orders df

df_ords.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'), index=False)
