==> Instacart Basket Analysis/03 Scripts/2.1 IC Combining Orders and Orders_products_prior Data Sets.ipynb <==
# 2.1 Combining Orders and Orders_products_prior Data Sets

## Contents
### Import libraries and data sets
### Consistency checks on ords_priors df
### Merge ords and ords_prior
### Export to pickle

## Import libraries and data sets

import pandas as pd
import numpy as np
import os

# create path
path = r'/Users/susanwang/Documents/CF_Tasks/Instacart Basket Analysis'

# import orders df
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'), index_col=False)

# check the output
df_ords.head()

# import order_products_prior df
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col=False)

# check the output
df_ords_prior.head()

## Consistency check on ords_prior df

# mixed_data check
for col in df_ords_prior.columns.to_list():
    weird = (df_ords_prior[[col]].map(type) != df_ords_prior[[col]].iloc[0].apply(type)).any(axis=1)
    if len(df_ords_prior[weird]) > 0:
        print(col)
    else:
        print(f"all same type for {col}")

# check for missing values
df_ords_prior.isnull().sum()

# check for duplicates
df_ords_prior_dups = df_ords_prior[df_ords_prior.duplicated()]

df_ords_prior_dups

## Merge ords and ords_prior

### Check dimensions of dfs

df_ords.shape

df_ords_prior.shape

### Merge on df_ords

# df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

I have commented out this version above with the indicator , because I need to do a second merge later and cannot have a _merge column in my dataset. Will create a df merge without indicator.

# df_merged_large.head()

# check indicator
# df_merged_large['_merge'].value_counts()

# df merge without indicator
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id')

df_merged_large.head()

df_merged_large.shape

## Export to pickle

df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
