==> Instacart Basket Analysis/03 Scripts/2.2 IC Combine Products with Orders.ipynb <==
# 2.2 Combine Products with Orders Data Sets

## Contents
### Import libraries and data sets
### Check shape of dfs before merge
### Merge products and orders_products_combined dfs
### Export merged df

## Import libraries and data sets

import pandas as pd
import numpy as np
import os

# path to project folder
path = r'/Users/susanwang/Documents/CF_Tasks/Instacart Basket Analysis'

# create df of products data
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col=False)

df_prods.head()

# create df of orders_products_combined
df_comb = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

df_comb.head()

## Check shape of dfs before merge

df_prods.shape

df_comb.shape

Common variable in both data sets is 'product_id'.

## Merge products and orders_products_combined dfs

# merge on outer to test
df_outer_test = df_comb.merge(df_prods, on='product_id', how='outer', indicator=True)

# check indicator
df_outer_test['_merge'].value_counts()

# merge on inner
df_ords_prods = df_comb.merge(df_prods, on='product_id', indicator=True)

df_ords_prods.head()

#check indicator for this merge
merge_counts = df_ords_prods['_merge'].value_counts()

merge_counts

# copy to excel report
merge_counts.to_clipboard()

# check df shape after merge
df_ords_prods.shape

## Export merged df

df_ords_prods.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge.pkl'))
