==> Instacart Basket Analysis/03 Scripts/3.1 IC Deriving New Variables.ipynb <==
# 3.1 IC Deriving New Variables

## Contents
### Import libraries and data sets
### Create subset with 1 million rows
### User defined functions
### Using loc() function
### Loc() derivation on entire data set
### If statements with for loops
### Deriving the "busiest_days" variable
### Identify busiest hours of the day
### Export as pickle

## Import libraries and data sets

import pandas as pd
import numpy as np
import os

# path to project folder
path = r'/Users/susanwang/Documents/CF_Tasks/Instacart Basket Analysis'

# import ords_prods_merge pickle
ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge.pkl'))

# check data
ords_prods_merge.head()

ords_prods_merge.shape

## Create a subset with 1 million rows

df = ords_prods_merge[:1000000]

df.shape

## User-Defined Functions

### We want to sort products into price ranges by flagging them.

# define function
def price_label(row):
    if row['prices'] <= 5:
        return 'Low-range product'
    elif (row['prices'] > 5) and (row['prices'] <= 15):
        return 'Mid-range product'
    elif row['prices'] > 15:
        return 'High-range product'
    else:
        return 'Not enough data'

df['price_range'] = df.apply(price_label, axis = 1)

# check value counts
df['price_range'].value_counts(dropna=False)

## Using loc() Function

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'


df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna=False)

## Loc() Derivation on entire dataset

ords_prods_merge.loc[ords_prods_merge['prices'] <= 5, 'price_range'] = 'Low-range product'

ords_prods_merge.loc[(ords_prods_merge['prices'] <= 15) & (ords_prods_merge['prices'] > 5), 'price_range'] = 'Mid-range product'

ords_prods_merge.loc[ords_prods_merge['prices'] > 15, 'price_range'] = 'High-range product'

# check counts
price_count = ords_prods_merge['price_range'].value_counts(dropna=False)

price_count

# copy frequency table to excel report
price_count.to_clipboard()

## If Statements with For Loops

### We want to summarize how busy each day of the week is.

# view frequency
ords_prods_merge['orders_day_of_the_week'].value_counts(dropna=False)

# loop through column and append to a list
result = []

for value in ords_prods_merge["orders_day_of_the_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

# set a new column to list
ords_prods_merge['busiest_day'] = result

# check value counts
bus_day_check = ords_prods_merge['busiest_day'].value_counts(dropna=False)

bus_day_check

# copy frequency table to excel report
bus_day_check.to_clipboard()

## Deriving the "Busiest days" variable

# view frequency of days in the df
ords_prods_merge['orders_day_of_the_week'].value_counts(dropna=False)

# create list
busiest = []

for value in ords_prods_merge["orders_day_of_the_week"]:
  if value <= 1:
    busiest.append("Busiest days")
  elif value == 3:
    busiest.append("Slowest days")
  elif value == 4:
    busiest.append("Slowest days")
  else:
    busiest.append("Regular days")

# set list as new column
ords_prods_merge['busiest_days'] = busiest

# check value counts of new column
check_days = ords_prods_merge['busiest_days'].value_counts(dropna=False)

check_days

# copy to excel report
check_days.to_clipboard()

# compare to original column
ords_prods_merge['orders_day_of_the_week'].value_counts(dropna=False)

The results look accurate. The two busiest days have almost as many orders as three regular days.

## Identify busiest hours of the day

# view order_hour_of_day frequencies
ords_prods_merge['order_hour_of_day'].value_counts(dropna=False)

### I will divide the hours of day into three periods. The 'Most orders' will be the top 8 hours by frequency. 'Average orders' will be the middle 8 hours. 'Fewest orders' will be the last third of the hours.

# create lists of hours in each category
high = [10, 11, 14, 15, 13, 12, 16, 9]
mid = [17, 8, 18, 19, 20, 7, 21, 22]
low = [23, 6, 0, 1, 5, 2, 4, 3]

print(len(high))
print(len(mid))
print(len(low))

### Use a subset to practice and test

sub_df = ords_prods_merge[:1000]

sub_df.loc[sub_df['order_hour_of_day'].isin(high), 'busiest_period_of_day'] = 'Most orders'

sub_df.loc[sub_df['order_hour_of_day'].isin(mid), 'busiest_period_of_day'] = 'Average orders'

sub_df.loc[sub_df['order_hour_of_day'].isin(low), 'busiest_period_of_day'] = 'Fewest orders'

sub_df['busiest_period_of_day'].value_counts(dropna=False)

### Apply to whole data set

ords_prods_merge.loc[ords_prods_merge['order_hour_of_day'].isin(high), 'busiest_period_of_day'] = 'Most orders'

ords_prods_merge.loc[ords_prods_merge['order_hour_of_day'].isin(mid), 'busiest_period_of_day'] = 'Average orders'

ords_prods_merge.loc[ords_prods_merge['order_hour_of_day'].isin(low), 'busiest_period_of_day'] = 'Fewest orders'

# check value counts and compare
check_hrs = ords_prods_merge['busiest_period_of_day'].value_counts(dropna=False)

check_hrs

# copy to excel report
check_hrs.to_clipboard()

ords_prods_merge.shape

# view head to check new columns
ords_prods_merge.head()

## Export as pickle

ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_derive.pkl'))
