==> Instacart Basket Analysis/03 Scripts/3.2 IC Grouping and Aggregating Data.ipynb <==
# 3.2 IC Grouping and Aggregating Data

## Contents
### Import libraries and data sets
### Create loyalty flag for existing customers
### Spending habits of three types of customers
### Create spending flag
### Create frequency flag
### Handling Outliers
### Export to pickle

## Import libraries and data sets

import pandas as pd
import numpy as np
import os

# path to project folder
path = r'/Users/susanwang/Documents/CF_Tasks/Instacart Basket Analysis'

# create df of ords_prods_derive
ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_derive.pkl'))

ords_prods_merge.head()

ords_prods_merge.shape

## Create a loyalty flag for existing customers

# create new column max_order for maximum number of orders for each customer
# note: error message upon use of np.max. suggestion to pass the string 'max' instead.
ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform('max')

# change display options to view all 100 rows 
pd.options.display.max_rows = None

ords_prods_merge[['user_id', 'order_number', 'max_order']].head(100)

# new column with loyalty flag
ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'
ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'
ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# check new column frequency
check_loyalty = ords_prods_merge['loyalty_flag'].value_counts(dropna=False)

check_loyalty

# copy to excel report
check_loyalty.to_clipboard()

# view sample (head is always the same user)
ords_prods_merge[['user_id', 'loyalty_flag', 'max_order']].sample(10)

## Spending habits of the three types of customers

### Basic statistics of the product prices

# groupby types of customers
loyalty_vs_price = ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean']})

loyalty_vs_price

# copy to report
loyalty_vs_price.to_clipboard()

It seems the most loyal customers spend on average less than the new or regular customers.

## Create spending flag based on average price across all their orders

Criteria: If the mean of the prices of products purchased by a user is lower than 10, then flag them as a “Low spender.” If the mean of the prices of products purchased by a user is higher than or equal to 10, then flag them as a “High spender.”

# aggregate and derive column 'avg_price' using transform
ords_prods_merge['avg_price'] = ords_prods_merge.groupby('user_id')['prices'].transform('mean')

# view and check head
ords_prods_merge[['user_id', 'prices', 'avg_price']].head(100)

# create spending flag
ords_prods_merge.loc[ords_prods_merge['avg_price'] < 10, 'spending_flag'] = 'Low spender'
ords_prods_merge.loc[ords_prods_merge['avg_price'] >= 10, 'spending_flag'] = 'High spender'

# check counts
check_spending = ords_prods_merge['spending_flag'].value_counts(dropna=False)

check_spending

# copy to excel report
check_spending.to_clipboard()

# check sample
ords_prods_merge[['user_id', 'avg_price', 'spending_flag']].sample(10)

ords_prods_merge.shape

## Create frequency flag for each customer based on days_since_prior_order column

If the median of “days_since_prior_order” is higher than 20, then the customer should be labeled a “Non-frequent customer.”
If the median is higher than 10 and lower than or equal to 20, then the customer should be labeled a “Regular customer.”
If the median is lower than or equal to 10, then the customer should be labeled a “Frequent customer.”

# create median_days column 
ords_prods_merge['median_days'] = ords_prods_merge.groupby('user_id')['days_since_prior_order'].transform('median')

ords_prods_merge[['user_id', 'days_since_prior_order', 'median_days']].head(100)

# create frequency_flag
ords_prods_merge.loc[ords_prods_merge['median_days'] > 20, 'frequency_flag'] = 'Non-frequent customer'
ords_prods_merge.loc[(ords_prods_merge['median_days'] > 10) & (ords_prods_merge['median_days'] <= 20), 'frequency_flag'] = 'Regular customer'
ords_prods_merge.loc[ords_prods_merge['median_days'] <= 10, 'frequency_flag'] = 'Frequent customer'

# check value counts
freq_counts = ords_prods_merge['frequency_flag'].value_counts(dropna=False)

freq_counts

# check NaN
ords_prods_merge[ords_prods_merge['frequency_flag'].isnull()]

The NaN results are due to the 5 customers who didn't make more than 1 order.

# copy frequency table to excel report
freq_counts.to_clipboard()

# check sample
ords_prods_merge[['user_id', 'median_days', 'frequency_flag']].sample(10)

## Handling Outliers (from Exercise 4.9)

# check prices max
ords_prods_merge['prices'].max()

# turn outliers into nan
ords_prods_merge.loc[ords_prods_merge['prices'] > 100, 'prices'] = np.nan

# check max again
ords_prods_merge['prices'].max()

## Export to pickle

# check shape
ords_prods_merge.shape

ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_agg.pkl'))
