==> Instacart Basket Analysis/03 Scripts/3.3 IC Customer Data Preparation and Merge.ipynb <==
# 3.3 IC Customer Data Preparation and Merge

## Contents
### Import libraries and customer data set
### Import customer data set with column restrictions
### Wrangling Procedures
### Data consistency checks
### Export cleaned data set
### Merge cleaned data set with ords_prods data set
### Crosstabs for NaN Values
### Test merge
### Merge into df
### Export merged df

## Import libraries and customer data set

import pandas as pd
import numpy as np
import os

# path to project folder
path = r'/Users/susanwang/Documents/CF_Tasks/Instacart Basket Analysis'

Take a preliminary look at the data set, import only first 1000 rows.

df_check = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'), index_col=False, nrows=1000)

df_check.shape

df_check.head()

The 'date_joined' may also not be necessary, as the 'loyalty_flag' will address the customer's status. 

For consistencies sake, all the columns should be in lower case, and 'Surnam' would be better named 'last_name'.

## Import customer data set with column restrictions

cust_list = ['user_id', 'First Name', 'Surnam', 'Gender', 'STATE', 'Age', 'n_dependants', 'fam_status', 'income']

df_cust = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'), usecols=cust_list, index_col=False)

df_cust.head()

## Wrangling Procedures

### Rename columns

df_cust.rename(columns={'First Name': 'first_name', 'Surnam': 'last_name', 'Gender': 'gender', 'STATE': 'state','Age': 'age'}, inplace=True)

#check column names
df_cust.head()

### Explore data

# shape of data
df_cust.shape

# tail
df_cust.tail()

# info
df_cust.info()

- Seems like there may be missing values in the first_name column.
- User_id should not be integer.

### Changing data type for user_id

df_cust['user_id'] = df_cust['user_id'].astype('str')

# check dtypee
df_cust['user_id'].dtype

### Descriptive stats

df_cust.describe()

## Data consistency checks

# check for mixed_type data
for col in df_cust.columns.to_list():
    weird = (df_cust[[col]].map(type) != df_cust[[col]].iloc[0].apply(type)).any(axis=1)
    if len(df_cust[weird]) > 0:
        print(col)
    else:
        print(f'all consistent in {col}')

# check for missing values
df_cust.isnull().sum()

df_cust[df_cust['first_name'].isnull() == True]

There are many NaN values in the first_name column. As it won't affect the analysis results, nor will I be using the column for grouping or aggregating purposes, I can leave the NaN values as is, or even drop the entire column.

# drop first_name column
df_cust = df_cust.drop(columns=['first_name'])

# check head
df_cust.head()

# check for missing values again
df_cust.isnull().sum()

# check for duplicates
cust_dups = df_cust[df_cust.duplicated()]

cust_dups

No duplicates found.

# check shape again
df_cust.shape

## Export cleaned data set to Prepared Data folder

df_cust.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'cust_clean.pkl'))

## Merge cleaned customer data with the ords_prods data set

# import ords_prods_agg data set
df_ords_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_agg.pkl'))

# check data
df_ords_prods.head()

df_ords_prods.shape

## Crosstabs for NaN Values in 'days_since_prior_order'

# create a crosstab 
crosstab = pd.crosstab(df_ords_prods['days_since_prior_order'], df_ords_prods['order_number'], dropna=False)

# save result to clipboard (and then paste into Excel)
crosstab.to_clipboard()

# check columns
df_ords_prods.columns

- Common key value is user_id.
- _merge flag column must be removed before making another merge, or else I cannot use the indicator flag.
- I will drop 'first_order' because it is also not needed for analysis.

# drop _merge column
df_ords_prods = df_ords_prods.drop(columns=['_merge', 'first_order'])

# check head
df_ords_prods.head()

## Test merge

# pd.merge(df_cust, df_ords_prods, on='user_id', indicator=True)

I get a warning that I am merging on integer and object, which means one of my user_id is not the right data type.

# check data types
df_cust['user_id'].dtype

df_ords_prods['user_id'].dtype

# change data type for df_ords_prods
df_ords_prods['user_id'] = df_ords_prods['user_id'].astype('str')

# check again
df_ords_prods['user_id'].dtype

pd.merge(df_cust, df_ords_prods, on='user_id', indicator=True)

## Merge into df

ords_prods_all = df_cust.merge(df_ords_prods, on='user_id', indicator=True)

# check value counts for _merge
ords_prods_all['_merge'].value_counts(dropna=False)

# check shape
ords_prods_all.shape

## Export merged df

ords_prods_all.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_all.pkl'))
